Changed some file locations: Recipe API search and show files moved to src/components/pages/section-nutrition. Changed element folder into circleEl to hold only my sad circle collection. Added pages.css - empty - to stand in for styling in pages. Should probably have individual css files to segment project further and avoid conflicts.
