+++ client/src/App.js
@@ -1,9 +1,9 @@
-import LoginForm from './components/Form/LoginForm';
+import LoginForm from './components/forms/LoginForm';
 //import Page from './components/Page';
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import Login from './components/pages/Login';
-import RegisterForm from './components/Form/RegisterForm';
-import { RecipeForm } from './components/Form/RecipeForm';
+import RegisterForm from './components/forms/RegisterForm';
+import { RecipeForm } from './components/forms/RecipeForm';
 
 function App() {
     return (
